fix(ui): blank dark screen - Issue due to Tldraw Server side render problems
