+++ components/canvas/Whiteboard.tsx
@@ -1,8 +1,12 @@
 'use client';
 
-import { Tldraw, Editor, getSnapshot, loadSnapshot } from '@tldraw/tldraw';
-import '@tldraw/tldraw/tldraw.css';
-import { useCallback, useEffect, useState, useRef } from 'react';
+import dynamic from 'next/dynamic';
+
+// Dynamically import the actual whiteboard component with SSR disabled
+const DynamicWhiteboard = dynamic(
+    () => import('./WhiteboardClient'),
+    { ssr: false, loading: () => <div className="flex items-center justify-center w-full h-full"><div className="w-8 h-8 border-4 border-primary border-t-transparent rounded-full animate-spin" /></div> }
+);
 
 interface WhiteboardProps {
     initialData?: string;
@@ -10,55 +14,6 @@ interface WhiteboardProps {
     readOnly?: boolean;
 }
 
-export function Whiteboard({ initialData, onSave, readOnly = false }: WhiteboardProps) {
-    const [editor, setEditor] = useState<Editor | null>(null);
-    const isInitialLoadRef = useRef(true);
-
-    const handleMount = useCallback((editor: Editor) => {
-        setEditor(editor);
-
-        if (initialData) {
-            try {
-                const snapshot = JSON.parse(initialData);
-                loadSnapshot(editor.store, snapshot);
-            } catch (error) {
-                console.error("Failed to load whiteboard data", error);
-            }
-        }
-    }, [initialData]);
-
-    // Track changes for saving
-    useEffect(() => {
-        if (!editor || readOnly || !onSave) return;
-
-        // Debounced save
-        let timeoutId: NodeJS.Timeout;
-
-        const cleanup = editor.store.listen(() => {
-            if (isInitialLoadRef.current) {
-                isInitialLoadRef.current = false;
-                return;
-            }
-
-            clearTimeout(timeoutId);
-            timeoutId = setTimeout(() => {
-                const snapshot = getSnapshot(editor.store);
-                onSave(JSON.stringify(snapshot));
-            }, 1000); // Save after 1 second of inactivity
-        });
-
-        return () => {
-            cleanup();
-            clearTimeout(timeoutId);
-        };
-    }, [editor, readOnly, onSave]);
-
-    return (
-        <div className="w-full h-full relative" style={{ zIndex: 10 }}>
-            <Tldraw 
-                onMount={handleMount}
-                hideUi={readOnly} // Hide UI tools if in readOnly mode
-            />
-        </div>
-    );
+export function Whiteboard(props: WhiteboardProps) {
+    return <DynamicWhiteboard {...props} />;
 }
